timer works, fixed pointer, timer color

+++ Assets/Scripts/Level.cs
@@ -8,7 +8,7 @@ public class Level : ScriptableObject {
     [SerializeField] public string description = "One of the levels of time";
     
     /** Amount of time to complete the level, in seconds. */
-    [SerializeField] public double time = 300;
+    [SerializeField] public int time = 300;
     /** Points needed to complete the level **/
     [SerializeField] public int scoreGoal = 2000;
